refatora objetos herdando da classe Desenhável
