+++ versao_final/Models/Desenhavel.py
@@ -3,38 +3,62 @@ import pygame as pg
 from pygame.sprite import Sprite
 from pygame.image import load
 
+
 class Desenhavel(Sprite, ABC):
-  @abstractmethod
-  def __init__(self, image_path: str, position : tuple = (0,0)):
-    super().__init__()
-    self.__image = load(image_path)
-    self.__rect = self.__image.get_rect(center=position)
-    self.__mask = pg.mask.from_surface(self.image)
-    self.__position = position
+    @abstractmethod
+    def __init__(self, image_path: str, position: tuple = (0, 0), size=None):
+        super().__init__()
+        loaded_image = load(image_path)
+        self.__image = pg.transform.scale(
+            loaded_image, (size, size)) if size else loaded_image
+        self.__rect = self.__image.get_rect(topleft=position)
+        self.__mask = pg.mask.from_surface(self.image)
+        self.__position = position
 
-  @property
-  def image(self):
-    return self.__image
+    @property
+    def image(self):
+        return self.__image
 
-  @property
-  def position(self):
-    return self.__position
+    @image.setter
+    def image(self, image):
+        self.__image = image
 
-  @property
-  def mask(self):
-    return self.__mask
+    @property
+    def rect(self):
+        return self.__rect
 
-  @property
-  def rect(self): return self.__rect 
-  
-  def set_rect_left(self, valor): self.__rect.left = valor
-  
-  def set_rect_right(self, valor): self.__rect.right = valor
+    @rect.setter
+    def rect(self, rect):
+        self.__rect = rect
 
-  def set_rect_top(self, valor): self.__rect.top = valor
+    @property
+    def position(self):
+        return self.__position
 
-  def set_rect_bottom(self, valor): self.__rect.bottom = valor
+    @property
+    def mask(self):
+        return self.__mask
 
-  @image.setter
-  def image(self, image):
-    self.__image = image
+    def set_rect_left(self, valor): self.__rect.left = valor
+
+    def set_rect_right(self, valor): self.__rect.right = valor
+
+    def set_rect_top(self, valor): self.__rect.top = valor
+
+    def set_rect_bottom(self, valor): self.__rect.bottom = valor
+
+    def get_rect_left(self):
+        return self.__rect.left
+
+    def get_rect_right(self):
+        return self.__rect.right
+
+    def get_rect_top(self):
+        return self.__rect.top
+
+    def get_rect_bottom(self):
+        return self.__rect.bottom
+
+    def update(self, deslocamento_x, deslocamento_y):
+        self.__rect.x += deslocamento_x
+        self.__rect.y += deslocamento_y
